Move util/dom and util/document
