--- src/client/client-util/index.js
@@ -1,1 +1,0 @@
-export * from './file';
