+++ src/util/index.js
@@ -2,10 +2,8 @@ export * from './obj';
 export * from './promise';
 export * from './is';
 export * from './assert';
-export * from './dom';
 export * from './cy';
 export * from './memoize';
 export * from './strings';
 export * from './cache';
 export * from './fetch';
-export * from './document';
